requirements.txt
about-time==4.2.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aiosqlite==0.21.0
alive-progress==3.2.0
altair==5.5.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
asyncua==1.1.5
attrs==25.1.0
autograd==1.7.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cma==3.2.2
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
datumaro==1.9.1
debugpy==1.8.12
decorator==5.2.0
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.9
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.17.0
Flask==3.1.0
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.2.0
geti-sdk==2.7.0
gevent==24.11.1
geventhttpclient==2.0.2
gitdb==4.0.12
GitPython==3.1.44
grapheme==0.6.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
imageio-ffmpeg==0.6.0
imagesize==1.4.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json-stream==2.3.3
json-stream-rs-tokenizer==0.4.27
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jstyleson==0.0.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lap==0.5.12
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
mpmath==1.3.0
multidict==6.1.0
narwhals==1.27.1
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
nibabel==5.3.2
ninja==1.11.1.3
nltk==3.9.1
nncf==2.15.0
notebook_shim==0.2.4
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.10.0.84
openvino==2024.6.0
openvino-dev==2024.6.0
openvino-telemetry==2025.0.1
openvino_model_api==0.2.5.1
orjson==3.10.7
overrides==7.7.0
ovmsclient==2023.1
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathvalidate==3.2.3
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
portalocker==3.1.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pycocotools==2.0.8
pycparser==2.22
pydeck==0.9.1
pydot==2.0.0
pyemd==1.0.0
Pygments==2.19.1
pymoo==0.6.1.3
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-rapidjson==1.20
pytz==2025.1
PyYAML==6.0.1
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semver==3.0.4
Send2Trash==1.8.3
shapely==2.0.7
simplejson==3.19.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
streamlit==1.42.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboardX==2.6.2.2
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.78
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3==1.26.20
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
yarl==1.18.3
yt-dlp==2025.2.19
zope.event==5.0
zope.interface==7.2